Added Unit tests for MetadataSchema Extension

+++ src/app/CHAOS.MCM/Extension/AMcmExtension.cs
@@ -95,7 +95,7 @@
         public bool HasPermissionToObject(ICallContext callContext, Guid objectGuid, FolderPermission permissions)
 	    {
             //review: look into using the folder returned from the database directly for the permission check
-            var folders    = McmRepository.FolderGet(null, objectGuid).Select(item => PermissionManager.GetFolders((uint)item.ID));
+            var folders    = McmRepository.FolderGet(null, objectGuid).Select(item => PermissionManager.GetFolders(item.ID));
             var userGuid   = callContext.User.Guid;
 			var groupGuids = callContext.Groups.Select( item => item.Guid );
